+++ app/page.tsx
@@ -11,7 +11,7 @@ import { TrendingFlatIcon } from '@/components/icons'
 import Background from '@/components/ui/background'
 import Input from '@/components/ui/input'
 import LinkButton from '@/components/ui/link-button'
-import { useSocket } from '@/hooks/use-socket'
+import { useSocket } from '@/hooks/socket/core/use-socket'
 import { isValidId } from '@/lib/utils/generate-game-id'
 import backgroundDesktopImage from '@/public/images/background-desktop-1.png'
 import backgroundMobileImage from '@/public/images/background-mobile-1.png'
