+++ app/result/[gameId]/page.tsx
@@ -7,8 +7,8 @@ import Image from 'next/image'
 import LoadingPage from '@/app/loading'
 import LinkButton from '@/components/ui/link-button'
 import { SITE_URL } from '@/constants/config'
-import { useSocket } from '@/hooks/use-socket'
-import { useSocketNavigation } from '@/hooks/use-socket-navigation'
+import { useSocket } from '@/hooks/socket/core/use-socket'
+import { useSocketNavigation } from '@/hooks/socket/policy/use-socket-navigation'
 import ConfettiComponent from '@/lib/confetti'
 import coin from '@/public/images/coin.png'
 import useGameStore from '@/store/game'
