+++ app/waiting/[gameId]/page.tsx
@@ -11,8 +11,8 @@ import PlayerGrid from '@/components/features/waiting/player-grid'
 import Background from '@/components/ui/background'
 import Button from '@/components/ui/button'
 import { useGameState } from '@/hooks/game/use-game-state'
-import { useSocket } from '@/hooks/use-socket'
-import { useSocketNavigation } from '@/hooks/use-socket-navigation'
+import { useSocket } from '@/hooks/socket/core/use-socket'
+import { useSocketNavigation } from '@/hooks/socket/policy/use-socket-navigation'
 import backgroundDesktopImage from '@/public/images/background-desktop-2.png'
 import backgroundMobileImage from '@/public/images/background-mobile-2.png'
 import useGameStore from '@/store/game'
