+++ hooks/socket/core/use-socket.ts
@@ -1,3 +1,7 @@
+/**
+ * [Core Layer] SocketContext를 안전하게 사용하기 위한 커스텀 훅입니다.
+ * SocketProvider 하위에서 socket 인스턴스와 공통 상태에 접근할 수 있게 합니다.
+ */
 import { useContext } from 'react'
 
 import { SocketContext } from '@/components/provider/socket-provider'
